Update index.html title and favicon; refactor styles and components for improved layout and functionality
