+++ index.html
@@ -2,13 +2,12 @@
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
-    <link rel="icon" type="image/svg+xml" href="/vite.svg" />
+    <link rel="icon" type="image/svg+xml" href="/src/assets/logo.png" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Vite + React + TS</title>
+    <title>Optimisation d'Affectation</title>
   </head>
   <body>
     <div id="root"></div>
-    <!--suppress HtmlUnknownTarget -->
     <script type="module" src="/src/main.tsx"></script>
   </body>
 </html>
