+++ src/App.css
@@ -28,7 +28,7 @@
     padding: 20px;
 }
 .heading {
-    font-size: 3.3rem;
+    font-size: 4.3rem;
     margin-bottom: 2.5rem;
 }
 .text {
