+++ src/index.css
@@ -64,11 +64,11 @@ table tr td:not(:first-child) {
   align-items: center;
   padding: 20px;
 }
-.workspace {
-  padding: 40vh 35vw;
-  width: fit-content;
+.workspace-container {
   background-image: radial-gradient(circle, #ccc 1px, transparent 1px);
   background-size: 30px 30px;
+  padding: 40vh 35vw;
+  width: fit-content;
 }
 .debug {
   border: solid red 1px !important;
